ledger: move non-lookup methods into a storage package (#4830)
