+++ ledger/catchpointwriter.go
@@ -28,6 +28,7 @@ import (
 	"github.com/algorand/msgp/msgp"
 
 	"github.com/algorand/go-algorand/data/basics"
+	"github.com/algorand/go-algorand/ledger/store"
 	"github.com/algorand/go-algorand/protocol"
 )
 
@@ -130,12 +131,14 @@ func (chunk catchpointFileChunkV6) empty() bool {
 }
 
 func makeCatchpointWriter(ctx context.Context, filePath string, tx *sql.Tx, maxResourcesPerChunk int) (*catchpointWriter, error) {
-	totalAccounts, err := totalAccounts(ctx, tx)
+	arw := store.NewAccountsSQLReaderWriter(tx)
+
+	totalAccounts, err := arw.TotalAccounts(ctx)
 	if err != nil {
 		return nil, err
 	}
 
-	totalKVs, err := totalKVs(ctx, tx)
+	totalKVs, err := arw.TotalKVs(ctx)
 	if err != nil {
 		return nil, err
 	}
