+++ ledger/txtail_test.go
@@ -30,6 +30,7 @@ import (
 	"github.com/algorand/go-algorand/data/bookkeeping"
 	"github.com/algorand/go-algorand/data/transactions"
 	"github.com/algorand/go-algorand/ledger/ledgercore"
+	"github.com/algorand/go-algorand/ledger/store"
 	ledgertesting "github.com/algorand/go-algorand/ledger/testing"
 	"github.com/algorand/go-algorand/protocol"
 	"github.com/algorand/go-algorand/test/partitiontest"
@@ -154,6 +155,8 @@ func (t *txTailTestLedger) initialize(ts *testing.T, protoVersion protocol.Conse
 	tx, err := t.trackerDBs.Wdb.Handle.Begin()
 	require.NoError(ts, err)
 
+	arw := store.NewAccountsSQLReaderWriter(tx)
+
 	accts := ledgertesting.RandomAccounts(20, true)
 	proto := config.Consensus[protoVersion]
 	newDB := accountsInitTest(ts, tx, accts, protoVersion)
@@ -166,12 +169,12 @@ func (t *txTailTestLedger) initialize(ts *testing.T, protoVersion protocol.Conse
 	for i := startRound; i <= t.Latest(); i++ {
 		blk, err := t.Block(i)
 		require.NoError(ts, err)
-		tail, err := txTailRoundFromBlock(blk)
+		tail, err := store.TxTailRoundFromBlock(blk)
 		require.NoError(ts, err)
-		encoded, _ := tail.encode()
+		encoded, _ := tail.Encode()
 		roundData = append(roundData, encoded)
 	}
-	err = txtailNewRound(context.Background(), tx, startRound, roundData, 0)
+	err = arw.TxtailNewRound(context.Background(), startRound, roundData, 0)
 	require.NoError(ts, err)
 	tx.Commit()
 	return nil
